+++ MODSRC/vazkii/botania/common/block/tile/TileManaBeacon.java
@@ -25,14 +25,14 @@ public class TileManaBeacon extends TileMod {
 			if(redstoneSide > 0)
 				redstone = true;
 		}
-		
+
 		if(!redstone) {
 			float[] color = EntitySheep.fleeceColorTable[worldObj.getBlockMetadata(xCoord, yCoord, zCoord)];
-			
+
 			Botania.proxy.setWispFXDistanceLimit(false);
 			for(int i = 0; i < 3; i++)
-				Botania.proxy.wispFX(worldObj, xCoord + 0.5, yCoord, zCoord + 0.5, color[0], color[1], color[2], (float) Math.random() * 5 + 1F, (float) (Math.random() - 0.5F), 10F * (float) Math.sqrt((256F / (256F - (float) yCoord))), (float) (Math.random() - 0.5F));
-			
+				Botania.proxy.wispFX(worldObj, xCoord + 0.5, yCoord, zCoord + 0.5, color[0], color[1], color[2], (float) Math.random() * 5 + 1F, (float) (Math.random() - 0.5F), 10F * (float) Math.sqrt(256F / (256F - yCoord)), (float) (Math.random() - 0.5F));
+
 			for(int i = 0; i < 4; i++)
 				Botania.proxy.wispFX(worldObj, xCoord + 0.5, 256, zCoord + 0.5, color[0], color[1], color[2], (float) Math.random() * 15 + 8F, (float) (Math.random() - 0.5F) * 8F, 0F, (float) (Math.random() - 0.5F) * 8F);
 			Botania.proxy.setWispFXDistanceLimit(true);
